requirements.txt
streamlit~=1.27.2
requests~=2.31.0
numpy~=1.26.1
#diffusers~=0.21.4
Pillow~=10.1.0
#transformers
#accelerates
#torch~=1.13.1
